requirements.txt

asgiref==3.8.1
Django==5.0.7
djangorestframework==3.15.2
mysqlclient==2.2.6
pillow==11.0.0
PyMySQL==1.1.1
pywin32==308
sqlparse==0.5.1
tzdata==2024.2


